requirements.txt
pymysql
paramiko
sshtunnel
pytest==5.4.3
pytest-html
pytest-xdist
pytest-rerunfailures
pytest-sugar
pytest-ordering
allure-pytest
pysftp==0.2.8
requests
mock
docopt
selenium
docker
docker-compose
xlrd
xlutils
xlwt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
regex
bs4
testrail
pytest-testrail
opencv-python
scikit-image
flaky
mysql-connector-python
jumpssh
pymongo
pyautogui